pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spatch-modules"
version = "1.0.0"
description = "Custom spatial transcriptomics analysis modules for the SpatialData + Sopa platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SPATCH Team", email = "spatch@example.com"}
]
keywords = ["spatial-transcriptomics", "spatial-omics", "spatialdata", "sopa"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "spatialdata>=0.6",
    "spatialdata-io>=0.2",
    "spatialdata-plot>=0.3",
    "sopa>=2.1",
    "napari-spatialdata>=0.5",
    "scanpy>=1.10",
    "squidpy>=1.5",
    "anndata>=0.10",
    "numpy>=1.24",
    "pandas>=2.0",
    "geopandas>=0.14",
    "shapely>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "tifffile>=2023.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

segmentation = [
    "cellpose>=3.0",
]

annotation = [
    "tangram-sc>=1.0",
    "celltypist>=1.6",
]

visualization = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.0",
]

all = [
    "spatch-modules[dev,segmentation,annotation,visualization]",
]

[project.scripts]
spatch-modules = "spatch_modules.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/spatch-modules"
Documentation = "https://spatch-modules.readthedocs.io"
Repository = "https://github.com/yourusername/spatch-modules"
Issues = "https://github.com/yourusername/spatch-modules/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["spatch_modules*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
